# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Onyx Domain API V1",
    "version" : "1.0.9"
  },
  "servers" : [ {
    "url" : "http://onyxdomainapi.azurewebsites.net"
  } ],
  "paths" : {
...
    "/Securities/List" : {
      "get" : {
        "tags" : [ "Securities" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Securities/Detail" : {
      "get" : {
        "tags" : [ "Securities" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Securities/Delete" : {
      "post" : {
        "tags" : [ "Securities" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Securities/Save" : {
      "post" : {
        "tags" : [ "Securities" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SecurityData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SecurityData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SecurityData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
    "/ShareClasses/Save" : {
      "post" : {
        "tags" : [ "ShareClasses" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareClassData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareClassData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareClassData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/ShareClasses/Delete" : {
      "post" : {
        "tags" : [ "ShareClasses" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
    "/StockExchanges/Save" : {
      "post" : {
        "tags" : [ "StockExchanges" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StockExchangeData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StockExchangeData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StockExchangeData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/StockExchanges/Delete" : {
      "post" : {
        "tags" : [ "StockExchanges" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Context" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "nullable" : true
          },
          "success" : {
            "type" : "boolean"
          },
          "data" : {
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "data" : "",
          "success" : true,
          "message" : "message"
        }
      },
...
      "IdOnly" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
      "SecurityData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "positionImpactDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "daysToSettlement" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expiryDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "strike" : {
            "type" : "number",
            "format" : "double",
            "nullable" : true
          },
          "ticker" : {
            "type" : "string",
            "nullable" : true
          },
          "isin" : {
            "type" : "string",
            "nullable" : true
          },
          "securityClassId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "dailyAdjustment" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "adrUnderline" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "adrUnderlineQuantity" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "baseSecurityPrice" : {
            "type" : "boolean"
          },
          "importPrice" : {
            "type" : "boolean"
          },
          "providerCode" : {
            "type" : "string",
            "nullable" : true
          },
          "countryId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "currencyId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "stockExchangeId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "gicsSubIndustryId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "baseSecurity" : {
            "type" : "string",
            "nullable" : true
          },
          "localSecurity" : {
            "type" : "string",
            "nullable" : true
          },
          "accountId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "sectorId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
...
      "ShareClassData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "managementFee1" : {
            "type" : "number",
            "format" : "double"
          },
          "performanceFee" : {
            "type" : "number",
            "format" : "double"
          },
          "managementFee2" : {
            "type" : "number",
            "format" : "double"
          },
          "managementFee3" : {
            "type" : "number",
            "format" : "double"
          },
          "limit1" : {
            "type" : "number",
            "format" : "double"
          },
          "limit2" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "additionalProperties" : false
      },
...
      "StockExchangeData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "fee" : {
            "type" : "number",
            "format" : "double",
            "nullable" : true
          },
          "country" : {
            "type" : "string",
            "nullable" : true
          },
          "settlementPlace" : {
            "type" : "string",
            "nullable" : true
          },
          "countryId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "holidayType" : {
            "type" : "string",
            "nullable" : true
          },
          "holidayTypeId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}